+++ Templates/SharePoint-ADFS/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Revert the change on the SKU of Public IP addresses, to use again Basic SKU when possible.
 * Update DSC modules used to latest version available.
+* Replace DSC module xDnsServer 2.0.0 with DnsServerDsc 3.0.0.
 
 ## Enhancements & bug-fixes - Published in September 29, 2022